Phyloseq objects uploaded
These are the phyloseq objects created using the new metadata with columns for binned depression and anxiety. "parkinsons.RData" is the bare phyloseq object, "pakinsons_final.RData" is the non-rarefied phyloseq object with filtering used to remove samples with less than 100 reads and non-bacterial samples, and "parkinsons_rare.RData" is the fully rarefied phyloseq object.